split parser tests module into a separate file
